# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Assignment Guy API — contract-first specification
 * OpenAPI spec version: 0.3.0
 */
export type UserProfileRole = typeof UserProfileRole[keyof typeof UserProfileRole];


export const UserProfileRole = {
  student: 'student',
  contributor: 'contributor',
  trusted_contributor: 'trusted_contributor',
  admin: 'admin',
} as const;

export interface School {
  school_id: string;
  school_name: string;
  country?: string | null;
  created_at: string;
}

export interface Department {
  department_id: string;
  school_id: string;
  department_name: string;
}

export interface Course {
  course_id: string;
  department_id: string;
  course_code: string;
  course_name: string;
  level?: number | null;
  semester?: string | null;
}

export interface UserProfile {
  id: string;
  email: string;
  displayName: string;
  profilePhoto?: string | null;
  school?: School | null;
  department?: Department | null;
  level?: number | null;
  courses: Course[];
  contributionScore: number;
  role: UserProfileRole;
  hasCompletedOnboarding: boolean;
}

export interface UserProfileResponse {
  success: boolean;
  data: UserProfile;
}

export interface UpdateProfileRequest {
  email?: string;
  displayName: string;
  schoolId?: string | null;
  departmentId?: string | null;
  level?: number | null;
  courseIds?: string[];
}

export interface HealthStatus {
  status: string;
}

export interface SchoolListResponse {
  success: boolean;
  data: School[];
}

export interface DepartmentListResponse {
  success: boolean;
  data: Department[];
}

export interface CourseListResponse {
  success: boolean;
  data: Course[];
}

export interface AssignmentSummary {
  assignment_id: string;
  title: string;
  description?: string | null;
  due_date?: string | null;
  assignment_type: string;
  status: string;
  verification_status: string;
  discussion_count: number;
  created_at: string;
  updated_at: string;
  course_id: string;
  course_name: string;
  course_code: string;
  uploader_display_name: string;
}

export interface AssignmentAttachmentItem {
  attachment_id: string;
  attachment_type: string;
  url?: string | null;
  filename?: string | null;
  file_size?: number | null;
  sort_order: number;
}

export interface AssignmentContextItem {
  context_id: string;
  section_title: string;
  content: string;
  sort_order: number;
}

export type AssignmentDetail = AssignmentSummary & {
  attachments: AssignmentAttachmentItem[];
  context: AssignmentContextItem[];
};

export interface CreateAssignmentRequest {
  courseId: string;
  title: string;
  description?: string;
  dueDate?: string;
  assignmentType?: string;
}

export interface AssignmentListResponse {
  success: boolean;
  data: AssignmentSummary[];
  total: number;
}

export interface AssignmentDetailResponse {
  success: boolean;
  data: AssignmentDetail;
}

export type CreateAssignmentResponseData = {
  assignment_id: string;
};

export interface CreateAssignmentResponse {
  success: boolean;
  data: CreateAssignmentResponseData;
}

export interface ErrorResponse {
  success: boolean;
  message: string;
  code: string;
}

export type ListSchoolsParams = {
/**
 * Filter schools by name (case-insensitive partial match)
 */
search?: string;
};

export type UpdateMyProfile200 = {
  success: boolean;
};

export type ListCoursesByDepartmentParams = {
/**
 * Filter by academic level (1=Year1, 2=Year2, 3=Year3, 4=Year4, 5=Postgrad, 0=Other)
 */
level?: number;
};

export type ListAssignmentsParams = {
/**
 * Filter by course ID
 */
courseId?: string;
/**
 * Partial title match (case-insensitive)
 */
search?: string;
/**
 * Assignment status filter: active | closed | archived
 */
status?: string;
/**
 * Max results (default 30, max 100)
 */
limit?: number;
/**
 * Pagination offset (default 0)
 */
offset?: number;
};